requirements.txt
coverage==4.2
docopt==0.6.2
funcsigs==1.0.2
mock==2.0.0
nose==1.3.7
pbr==1.10.0
Pillow==3.4.2
psycopg2==2.6.2
six==1.10.0
virtualenv==15.1.0
